Type saga continues; fixed sgemm ukernel signature.
Details:
- Changed double* pointers in sgemm function signature to float*. At
  this point I've lost track of whether this was my fault or another
  dormant bug like the one described in ece9f6a, but at this point I
  no longer care. It's one of those days (aka I didn't ask for this).

+++ kernels/knl/3/bli_sgemm_knl_asm_24x16.c
@@ -183,11 +183,11 @@ static int32_t offsets[32] __attribute__((aligned(64))) =
 void bli_sgemm_knl_asm_24x16
      (
        dim_t               k_,
-       double*    restrict alpha,
-       double*    restrict a,
-       double*    restrict b,
-       double*    restrict beta,
-       double*    restrict c, inc_t rs_c_, inc_t cs_c_,
+       float*     restrict alpha,
+       float*     restrict a,
+       float*     restrict b,
+       float*     restrict beta,
+       float*     restrict c, inc_t rs_c_, inc_t cs_c_,
        auxinfo_t* restrict data,
        cntx_t*    restrict cntx
      )
